chore: relaunch repo baseline (LICENSE, layout, CI/CD, docs)
Add MIT LICENSE and AGENTS.md, expand README with install/usage/dev
instructions. Restructure into standard cmd/ + internal/ Go layout,
translate the last Chinese code comment to English (test fixtures with
intentional CJK content are left untouched — they cover multi-byte
UTF-8 handling). Fix golangci-lint findings (unchecked errors,
ineffectual assignment, tagged switch) and pin rules in .golangci.yml.

Replace the Azure DevOps pipeline with GitHub Actions: ci.yml runs
gofmt/vet/lint/test on every push and PR to main, release.yml builds
all six target binaries and publishes a GitHub Release only on tag
push (previously every push to main attempted a release). Drop the
now-obsolete Azure-specific scripts. Install the git-commit skill and
a context7 MCP config for AI-assisted development going forward.
