+++ cmd/nsis-formatter/expected/commontest.nsi
