+++ cmd/nsis-formatter/expected/commontest2.nsi
