+++ cmd/nsis-formatter/expected/commontest3.nsi
