+++ cmd/nsis-formatter/expected/commontest4.nsi
