+++ cmd/nsis-formatter/expected/commontest5.nsi
