+++ cmd/nsis-formatter/expected/empty-lines.nsi
