+++ cmd/nsis-formatter/expected/space-indentation-2.nsi
