+++ cmd/nsis-formatter/expected/space-indentation-3.nsi
