+++ cmd/nsis-formatter/expected/space-indentation.nsi
