+++ cmd/nsis-formatter/expected/tab-indentation-2.nsi
