+++ cmd/nsis-formatter/expected/tab-indentation.nsi
