+++ cmd/nsis-formatter/fixtures/commontest.nsi
