+++ cmd/nsis-formatter/fixtures/commontest2.nsi
