+++ cmd/nsis-formatter/fixtures/commontest3.nsi
