+++ cmd/nsis-formatter/fixtures/commontest4.nsi
