+++ cmd/nsis-formatter/fixtures/commontest5.nsi
