+++ cmd/nsis-formatter/fixtures/empty-lines.nsi
