+++ cmd/nsis-formatter/fixtures/indentation.nsi
